+++ CHANGELOG.md
@@ -8,13 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Changed
 
-- Temple Scanner now displays direction on the item itself, rather than printing coordinates in the chatbox.
+- Temple Scanner now displays direction on the item
 
 ### Contributors for this release
 
 - Vivian Zane, Dweblenod, kirderf1
-
-## [Unreleased]
 
 ### Changed